refactor: update TransactionCellDetail to TransactionCellScript
